Automatically convert files URL to hash URLs for IE8 and less
Since IE8 and less don't support the history API, we need to use
hashchange as a workaround. To make this seamless the server will
initially convert the URL to a hash URL in such cases.

In that specific case, the file list will be loaded only through ajax,
and will not be returned directly by index.php